Add a QueryHandler class and move to a new file.
